requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sentence-transformers
faiss-cpu
numpy
python-dotenv